--- src/types.ts
@@ -1,3 +1,0 @@
-export type Nullish<T> = T | null | undefined
-
-export type Callback<T> = () => T
