+++ tests/jest.config.dom.js
@@ -3,6 +3,8 @@ const baseConfig = require('kcd-scripts/jest')
 
 module.exports = {
   ...baseConfig,
+  moduleFileExtensions: [...baseConfig.moduleFileExtensions, 'd.ts'],
+  moduleDirectories: [...baseConfig.moduleDirectories, 'types'],
   rootDir: path.join(__dirname, '..'),
   displayName: 'dom',
   coveragePathIgnorePatterns: [
