+++ tests/jest.config.node.js
@@ -3,6 +3,8 @@ const baseConfig = require('kcd-scripts/jest')
 
 module.exports = {
   ...baseConfig,
+  moduleFileExtensions: [...baseConfig.moduleFileExtensions, 'd.ts'],
+  moduleDirectories: [...baseConfig.moduleDirectories, 'types'],
   rootDir: path.join(__dirname, '..'),
   displayName: 'node',
   testEnvironment: 'jest-environment-node',
